[development/gfx] fixing all compiler warnings in msvc

+++ source/1_engine/rex_renderer_core/include/rex_renderer_core/system/rhi.h
@@ -118,7 +118,7 @@ namespace rex
     void draw_indexed(s32 instanceCount, s32 startInstance, s32 indexCount, s32 startIndex, s32 baseVertexLoc);
 
     void reset_upload_buffer();
-    void update_buffer(const ResourceSlot& slot, const void* data, s32 size);
+    void update_buffer(const ResourceSlot& slot, const void* data, s64 size);
 
   }
 }
